Move mass calculation primitives to consensus-core
