+++ consensus/core/src/api/mod.rs
@@ -9,7 +9,7 @@ use crate::{
         block::{BlockProcessResult, RuleError},
         tx::TxResult,
     },
-    tx::{MutableTransaction, Transaction, TransactionOutput},
+    tx::{MutableTransaction, Transaction},
 };
 
 /// Abstracts the consensus external API
@@ -27,7 +27,6 @@ pub trait ConsensusApi: Send + Sync {
     fn validate_mempool_transaction_and_populate(self: Arc<Self>, transaction: &mut MutableTransaction) -> TxResult<()>;
 
     fn calculate_transaction_mass(self: Arc<Self>, transaction: &Transaction) -> u64;
-    fn calculate_transaction_output_estimated_serialized_size(self: Arc<Self>, transaction_output: &TransactionOutput) -> u64;
 
     fn get_virtual_daa_score(self: Arc<Self>) -> u64;
 }
