+++ consensus/src/processes/mass.rs
@@ -1,8 +1,4 @@
-use consensus_core::{
-    subnets::SUBNETWORK_ID_SIZE,
-    tx::{Transaction, TransactionInput, TransactionOutput},
-};
-use hashes::HASH_SIZE;
+use consensus_core::{mass::transaction_estimated_serialized_size, tx::Transaction};
 
 pub struct MassCalculator {
     mass_per_tx_byte: u64,
@@ -34,56 +30,4 @@ impl MassCalculator {
 
         mass_for_size + total_script_public_key_mass + total_sigops_mass
     }
-}
-
-// transaction_estimated_serialized_size is the estimated size of a transaction in some
-// serialization. This has to be deterministic, but not necessarily accurate, since
-// it's only used as the size component in the transaction and block mass limit
-// calculation.
-fn transaction_estimated_serialized_size(tx: &Transaction) -> u64 {
-    let mut size: u64 = 0;
-    size += 2; // Tx version (u16)
-    size += 8; // Number of inputs (u64)
-    let inputs_size: u64 = tx.inputs.iter().map(transaction_input_estimated_serialized_size).sum();
-    size += inputs_size;
-
-    size += 8; // number of outputs (u64)
-    let outputs_size: u64 = tx.outputs.iter().map(transaction_output_estimated_serialized_size).sum();
-    size += outputs_size;
-
-    size += 8; // lock time (u64)
-    size += SUBNETWORK_ID_SIZE as u64;
-    size += 8; // gas (u64)
-    size += HASH_SIZE as u64; // payload hash
-
-    size += 8; // length of the payload (u64)
-    size += tx.payload.len() as u64;
-    size
-}
-
-fn transaction_input_estimated_serialized_size(input: &TransactionInput) -> u64 {
-    let mut size = 0;
-    size += outpoint_estimated_serialized_size();
-
-    size += 8; // length of signature script (u64)
-    size += input.signature_script.len() as u64;
-
-    size += 8; // sequence (uint64)
-    size
-}
-
-const fn outpoint_estimated_serialized_size() -> u64 {
-    let mut size: u64 = 0;
-    size += HASH_SIZE as u64; // Previous tx ID
-    size += 4; // Index (u32)
-    size
-}
-
-pub fn transaction_output_estimated_serialized_size(output: &TransactionOutput) -> u64 {
-    let mut size: u64 = 0;
-    size += 8; // value (u64)
-    size += 2; // output.ScriptPublicKey.Version (u16)
-    size += 8; // length of script public key (u64)
-    size += output.script_public_key.script().len() as u64;
-    size
 }
